pyproject.toml
[project]
name = "open-deepagents-env"
version = "0.1.0"
description = "Local Python environment for experimenting with LangChain Deep Agents"
requires-python = ">=3.11,<3.13"
dependencies = [
    "deepagents",
    "langchain-openai",
    "python-dotenv",
    "tavily-python",
]

[tool.uv]
package = false
